+++ R/fp.logit.R
@@ -33,7 +33,7 @@
 #' fp.logit(obs, n, dose)
 #'
 #' @param obs Numeric vector of the number of patients experiencing the event of interest
-#'   at each dose level. Must be non-negative integers.
+#'   or normalized equivalent efficacy score at each dose level.
 #' @param n Numeric vector of the total number of patients treated at each dose level.
 #'   Must be positive integers. Length must match \code{obs} and \code{dose}.
 #' @param dose Numeric vector of dose levels investigated. Should be positive values
@@ -78,6 +78,7 @@
 #' from this function.
 #'
 #' @import mfp
+#' @importFrom stats quasi
 #' @keywords dose-response fractional-polynomial clinical-trials regression
 #' @export
 
@@ -85,13 +86,11 @@ fp.logit <- function(obs,n,dose)
 {
   prob    <- obs/n
   ld      <- length(prob)
-  obse.df <- NULL
-  for(i in 1:ld){
-    obse.df <- rbind(obse.df,data.frame(nEff=c(rep(1,obs[i]),rep(0,n[i]-obs[i])),dose=dose[i]))
-  }
-  fpfit <- suppressWarnings(mfp::mfp(factor(nEff)~fp(dose,df=4,select=0.99999,alpha=0.99999),family=binomial,data=obse.df))
+  obse.df <- data.frame(prob=prob,dose=dose)
+  fpfit <- suppressWarnings(mfp::mfp(prob~fp(dose,df=4,select=0.99999,alpha=0.99999),family=quasi(link="logit",variance="mu(1-mu)"),data=obse.df))
   fpnum <- apply(as.matrix(dose),1,function(x){min((1:sum(n))[obse.df$dose==x])})
   fp.obspe <- as.numeric((fpfit$fit$fitted.values)[fpnum])
   return(fp.obspe)
+
 }